+++ test/capacitors/HashChainCapacitor.t.sol
@@ -72,7 +72,7 @@ contract HashChainCapacitorTest is Setup {
 
         hashedRoot = keccak256(abi.encode(hashedRoot, _message_2));
 
-        bytes32 root = _hcCapacitor.getRootByCount(0);
+        bytes32 root = _hcCapacitor.getRootByCount(uint64(0));
         assertEq(root, hashedRoot);
 
         roots.push(_message_0);
@@ -114,13 +114,13 @@ contract HashChainCapacitorTest is Setup {
     }
 
     function _assertNextPacket(bytes32 root_, uint256 packetId_) private {
-        uint256 nextPacketId = _hcCapacitor.getLatestPacketCount() + 1;
+        uint64 nextPacketId = uint64(_hcCapacitor.getLatestPacketCount() + 1);
         bytes32 root = _hcCapacitor.getRootByCount(nextPacketId);
         assertEq(root, root_, "Root Invalid");
         assertEq(nextPacketId, packetId_, "packetId Invalid");
     }
 
-    function _assertPacketById(bytes32 root_, uint256 packetId_) private {
+    function _assertPacketById(bytes32 root_, uint64 packetId_) private {
         bytes32 root = _hcCapacitor.getRootByCount(packetId_);
         bytes32 packedRoot = root_;
         if (root != bytes32(0))
