+++ test/capacitors/SingleCapacitor.t.sol
@@ -110,13 +110,13 @@ contract SingleCapacitorTest is Test {
     }
 
     function _assertNextPacket(bytes32 root_, uint256 packetId_) private {
-        uint256 nextPacketId = _sa.getLatestPacketCount() + 1;
+        uint64 nextPacketId = uint64(_sa.getLatestPacketCount() + 1);
         bytes32 root = _sa.getRootByCount(nextPacketId);
         assertEq(root, root_, "Root Invalid");
         assertEq(nextPacketId, packetId_, "packetId Invalid");
     }
 
-    function _assertPacketById(bytes32 root_, uint256 packetId_) private {
+    function _assertPacketById(bytes32 root_, uint64 packetId_) private {
         bytes32 root = _sa.getRootByCount(packetId_);
         assertEq(root, root_, "Root Invalid");
     }
